+++ transforms/add-prop/__testfixtures__/Object.input.js
@@ -1,3 +1,0 @@
-import Button from "carbon-react/lib/components/button";
-
-export const asObject = () => <Button info={{prop: 1, prop: 2, prop:3}} />