+++ transforms/replace-row-column-with-grid/__testfixtures__/WithPropsPassedDown.input.js
@@ -1,0 +1,13 @@
+import { Row, Column } from "carbon-react/lib/components/row";
+
+const WithPropsPassedDown = (gutter, columns, otherProp) => {
+  return (
+    <Row {...{gutter, columns}}>
+      <Column {...{columns, otherProp}}>col1</Column>
+      <Column>col2</Column>
+      <Column>col3</Column>
+    </Row>
+    );
+};
+
+export default WithPropsPassedDown;